% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_load.R
\name{acs_load}
\alias{acs_load}
\title{acs_load}
\usage{
acs_load(geog, state, vars_load, year_val = 2019)
}
\arguments{
\item{geog}{a census geography; e.g., "place"}

\item{state}{a state; can enter NULL to get for nation}

\item{vars_load}{variables to load based on tidycensus variable list.}

\item{year_val}{a year to load data from}
}
\value{
ACS dataset
}
\description{
Function to load acs 5-year data for set of variables easily at given geography
}
