% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_load.R
\name{acspull}
\alias{acspull}
\title{Pull ACS Data}
\usage{
acspull(
  varlist,
  filename,
  year = 2020,
  geog = "block group",
  varscensus = NULL,
  state = "Maryland"
)
}
\arguments{
\item{varlist}{Vector of variables to download}

\item{filename}{String to save file as}

\item{year}{ACS5 year to download data from; default 2020}

\item{geog}{Geography of data to download; default block group}

\item{varscensus}{tidycensus downloaded variable names using tidycensus::load_variables; default NULL, which loads in ACS5 variables to merge}

\item{state}{tidycensus state; default "Maryland"}
}
\value{
Creates data directories to save loaded ACS5 data in
}
\description{
Function for downloading acs data at given geography and year
}
