% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_functs.R
\name{ancestry_recode}
\alias{ancestry_recode}
\title{Ancestry recode}
\usage{
ancestry_recode(df, ancestry_col)
}
\arguments{
\item{df}{ACS dataframe with ancestry values contained in one column}

\item{ancestry_col}{Name of ancestry column to recode}
}
\value{
Dataframe with recoded ancestry values
}
\description{
Ancestry recode
}
