% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_functs.R
\name{asian_country_recode}
\alias{asian_country_recode}
\title{Recode ACS Asian countries
Recodes Asian country names in ACS data, shortening India/China and two or more}
\usage{
asian_country_recode(df, ethnicity_col)
}
\arguments{
\item{df}{ACS dataframe}

\item{ethnicity_col}{Column to recode}
}
\value{
Dataframe with recoded Asian country names
}
\description{
Recode ACS Asian countries
Recodes Asian country names in ACS data, shortening India/China and two or more
}
