% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{base_process_vetagesex}
\alias{base_process_vetagesex}
\title{Base-process veteran-status by sex and age}
\usage{
base_process_vetagesex(df)
}
\arguments{
\item{df}{Tidycensus ACS5 dataframe on veteran status by age and sex}
}
\value{
Label-separated ACS5 dataframe on veteran status by age and sex
}
\description{
Conducts base-processing of veteran status by sex and age. This dataset is formatted oddly (separated-columns are somewhat inconsistent), so just separates label into sex, age, and veteran columns
}
