% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_functs.R
\name{birth_recode}
\alias{birth_recode}
\title{Birth recode
Function to recode birth values in ACS dataframe containing column for place of birth}
\usage{
birth_recode(df, birth_col)
}
\arguments{
\item{df}{Lightly-processed ACS dataframe with birth column}

\item{birth_col}{Name of birth column to recode}
}
\value{
Dataframe with additional column for place of birth
}
\description{
Birth recode
Function to recode birth values in ACS dataframe containing column for place of birth
}
