% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_formatter.R
\name{col_formatter}
\alias{col_formatter}
\title{Function to remove extraneous characters from strings}
\usage{
col_formatter(df)
}
\arguments{
\item{df}{a tidycensus dataframe}
}
\value{
a tidycensus dataframe with extraneous characters removed
}
\description{
Function to remove extraneous characters from strings
}
