% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_functs.R
\name{comp_recode}
\alias{comp_recode}
\title{Recode ACS computer/internet access variables}
\usage{
comp_recode(df, comp_col, int_col, new_comp = "comp_int")
}
\arguments{
\item{df}{A tidycensus ACS dataset with a column for computer access and one for internet access.}

\item{comp_col}{The name of the computer access column.}

\item{int_col}{The name of the internet accesss column.}

\item{new_comp}{The name of the new column}
}
\value{
A tidycensus ACS dataset with a factor-ordered computer/internet access column
}
\description{
Function to recode tidycensus ACS dataset with a column for computer access and a column for internet access to one column with shortened descriptors.
}
