% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_functs.R
\name{cost_burden_recode}
\alias{cost_burden_recode}
\title{Cost burden recode
Function to recode ACS cost burden values in one column}
\usage{
cost_burden_recode(df, cost_col)
}
\arguments{
\item{df}{Dataframe with ACS cost burden values}

\item{cost_col}{Name of column containing cost burden categories}
}
\value{
Dataframe with recoded cost burden values
}
\description{
Cost burden recode
Function to recode ACS cost burden values in one column
}
