% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_binder.R
\name{data_binder}
\alias{data_binder}
\title{Data binder}
\usage{
data_binder(file_root, dir_root)
}
\arguments{
\item{file_root}{root name of shared acs files want to bind together}

\item{dir_root}{directory path containing shared acs files}
}
\value{
dataframe
}
\description{
A function to load in a set of ACS rds files at different scales (e.g., place, county) together
}
