% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{data_puller}
\alias{data_puller}
\title{Data puller}
\usage{
data_puller(root_string, data_string)
}
\arguments{
\item{root_string}{Root of file name (e.g., municipalities or counties)}

\item{data_string}{Name of data directory (e.g., age by sex)}
}
\value{
Dataframe
}
\description{
Function to read in rds data with a common name from a data directory
}
