% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{data_saver}
\alias{data_saver}
\title{Save dataframe}
\usage{
data_saver(
  object,
  place_string,
  data_string,
  desc_year = "acs5_2019",
  place_dir = "all"
)
}
\arguments{
\item{object}{Dataframe to save as rds file}

\item{place_string}{String name of type of place in dataframe (e.g., municipalities)}

\item{data_string}{String name of}

\item{desc_year}{Year of data}

\item{place_dir}{Name of place directory}
}
\value{
Saved dataframe in created output directories associated with year and place data describes
}
\description{
Saves dataframe to output directories
}
