% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_pct_moe.R
\name{derive_pct_est_moe}
\alias{derive_pct_est_moe}
\title{Derive percent/ratio margin of error for groups}
\usage{
derive_pct_est_moe(
  df,
  proportion_col,
  aggregate_est,
  aggregate_moe,
  component_est = "estimate",
  component_moe = "moe",
  type_moe = "proportion"
)
}
\arguments{
\item{df}{A tidycensus ACS dataframe with rows representing component values and a column with a common group total.}

\item{proportion_col}{The name of the proportion column to create}

\item{aggregate_est}{A column representing an aggregate total for the group; the denominator in the proportion}

\item{aggregate_moe}{A column representing the margin of error for the aggregate estimate}

\item{component_est}{A column representing the value of the component value for which a proportion is being calculated; the numerator in the proportion}

\item{component_moe}{A column representing the marging of error of the component value}

\item{type_moe}{"proportion" or "ratio" for whether a proportion or ratio is being constructed}
}
\value{
A tidycensus dataframe with a calculated proportion or ratio and its margin of error.
}
\description{
A function to calculate the margin of error for proportions or ratios of tidycensus ACS dataframes, which are not included by default. Dataset should be structured with an aggregate column and margin of error uniform across rows, and a column for component values/margins of error. Formula construction follows ACS guidance in "Understanding and Using ACS Data," available here: https://www.census.gov/programs-surveys/acs/guidance/handbooks/general.html.
}
