% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disab_pull.R
\name{disab_pull}
\alias{disab_pull}
\title{Pull disability values from concept column}
\usage{
disab_pull(df, new_col = disability, concept_col = concept)
}
\arguments{
\item{df}{ACS dataset on disability-types, contained in one column}

\item{new_col}{Name of column with disability values pulled}

\item{concept_col}{Name of column with disability types}
}
\value{

}
\description{
Pull disability values from concept column
}
