% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{educ_race_process}
\alias{educ_race_process}
\title{Process educational attainment by race
Processes ACS5 tidycensus downloaded data on educational attainment by race}
\usage{
educ_race_process(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on educational attainment by race}
}
\value{
Processed data on educational attainment by race
}
\description{
Process educational attainment by race
Processes ACS5 tidycensus downloaded data on educational attainment by race
}
