% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educ_recode.R
\name{educ_recode}
\alias{educ_recode}
\title{Recode education values}
\usage{
educ_recode(df, educ_col, new_educ = "educ_recode")
}
\arguments{
\item{df}{A tidycensus dataframe with an education column}

\item{educ_col}{The column name of  to the education column in the tidycensus dataframe}

\item{new_educ}{The name of the new column containing the recoded education values; by edfault, educ_recode}
}
\value{
A dataframe with a recoded education column with shorter names
}
\description{
A function to recode tidycensus education columns to a shorter string for graphing purposes
}
