% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_moe_derive.R
\name{est_moe_derive}
\alias{est_moe_derive}
\title{Group ACS data and calculate derived-MOEs}
\usage{
est_moe_derive(
  df,
  group_cols,
  est_col = "estimate",
  moe_col = "moe",
  name_col = paste(group_cols, collapse = "_")
)
}
\arguments{
\item{df}{A tidy-census ACS dataframe.}

\item{group_cols}{Names of columns to group the dataframe by, formatted as a character vector.}

\item{est_col}{Name of estimate column, typically "estimate", formatted as a character string.}

\item{moe_col}{Name of margin of error column, typically "moe", formatted as a character string.}

\item{name_col}{Name of new, aggregated moe column, formatted as character string.}
}
\value{
A dataframe with estimates aggregated across group columns and a derived margin of error.
}
\description{
Groups tidy-census ACS dataframe by set of group columns, calculates aggregated estimates, and derives margin of error for aggregated estimates, per ACS instructions.
}
