% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_recode.R
\name{health_insur_recode}
\alias{health_insur_recode}
\title{Health insurance recode
Recodes ACS health insurance number of plans and type columns into one column, and creates factor column}
\usage{
health_insur_recode(df, num, type)
}
\arguments{
\item{df}{ACS dataframe with type of health insurance plan column and number of plans column}

\item{num}{Name of 'number of plans' column}

\item{type}{Name of type of plans column}
}
\value{
Dataframe with health-insurance categories recoded in one column
}
\description{
Health insurance recode
Recodes ACS health insurance number of plans and type columns into one column, and creates factor column
}
