% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_recode.R
\name{incpov_recode}
\alias{incpov_recode}
\title{Recode income/poverty
Recodes ACS dataframe income to poverty ratio column}
\usage{
incpov_recode(df, var)
}
\arguments{
\item{df}{Dataframe with income to poverty ratio column}

\item{var}{Name of income to poverty ratio column}
}
\value{
Dataframe with income to poverty ratio column recoded and sorted/factored
}
\description{
Recode income/poverty
Recodes ACS dataframe income to poverty ratio column
}
