% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_col.R
\name{location_col}
\alias{location_col}
\title{Create location column based on ACS dataframe name values}
\usage{
location_col(df)
}
\arguments{
\item{df}{Tidycensus acs dataframe with a name column containing place names for Takoma Park, Montgomery County, PG County, and/or Maryland}
}
\value{
A df with a location column with shortened version of location names
}
\description{
Create location column based on ACS dataframe name values
}
