% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_functs.R
\name{old_recode}
\alias{old_recode}
\title{Old recode
Function to recode ACS dataframe with age values greater than 75 to 75 and older}
\usage{
old_recode(df, age_col)
}
\arguments{
\item{df}{Dataframe with age column}

\item{age_col}{Name of age column}
}
\value{
Dataframe with recoded age column
}
\description{
Old recode
Function to recode ACS dataframe with age values greater than 75 to 75 and older
}
