% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_recode.R
\name{overall_check}
\alias{overall_check}
\title{Overall check
Function to check if vector has an overall value and to sort so that's last}
\usage{
overall_check(vec)
}
\arguments{
\item{vec}{Vector of values that might contain an "overall" value}
}
\value{
Ordered vector with "overall" moved to end if it's present
}
\description{
Overall check
Function to check if vector has an overall value and to sort so that's last
}
