% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{proces_family_age_tenure}
\alias{proces_family_age_tenure}
\title{Process Tenure by Family-type and Age}
\usage{
proces_family_age_tenure(df)
}
\arguments{
\item{df}{Processes ACS5 tidycensus downloaded data on tenure by family status and age}
}
\value{
Lightly processed data on tenure by family status and age
}
\description{
Processes ACS5 tidycensus downloaded data on tenure by family status and age
}
