% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_age_gender}
\alias{process_age_gender}
\title{Process age/sex data
Processes tidycensus downloaded ACS5 data on sex by age}
\usage{
process_age_gender(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on sex by age}
}
\value{
Processed data on age by sex
}
\description{
Process age/sex data
Processes tidycensus downloaded ACS5 data on sex by age
}
