% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_age_income}
\alias{process_age_income}
\title{Process age and income data}
\usage{
process_age_income(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on age and income in last 12 months}
}
\value{
Lightly processed data on age and income in last 12 months
}
\description{
Processes ACS5 tidycensus downloaded data on age and income in last 12 months
}
