% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_age_lang}
\alias{process_age_lang}
\title{Process age/english language proficiency}
\usage{
process_age_lang(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on age and English-language proficiency}
}
\value{
processed data on age/English language proficiency
}
\description{
Processes data on age and English language proficiency
}
