% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_age_overall}
\alias{process_age_overall}
\title{Process age data}
\usage{
process_age_overall(df)
}
\arguments{
\item{df}{Dataframe containing SEX BY AGE table, table B01001 in 2020 5-year ACS}
}
\value{
Dataframe grouped by age group at given geography
}
\description{
Processes tidycensus downloaded age data, table B01001
}
