% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_ancestry}
\alias{process_ancestry}
\title{Process ancestry data
Processes tidycensus downloaded ancestry data}
\usage{
process_ancestry(df, sub = F)
}
\arguments{
\item{df}{tidycensus downloaded ancestry data}

\item{sub}{Whether to calculate percentages of country for a given continent}
}
\value{
Processed tidycensus ancestry data
}
\description{
Process ancestry data
Processes tidycensus downloaded ancestry data
}
