% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_asian_disaggregated}
\alias{process_asian_disaggregated}
\title{Process Asian ethnicity
Processes tidycensus downloaded data on Asian ethnicity}
\usage{
process_asian_disaggregated(df)
}
\arguments{
\item{df}{tidycensus downloaded data on Asian by ethnicity}
}
\value{
processed data on Asian ethnicity
}
\description{
Process Asian ethnicity
Processes tidycensus downloaded data on Asian ethnicity
}
