% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_burden_owners}
\alias{process_burden_owners}
\title{Process cost-burden for homeowners
Processes ACS5 tidycensus downloaded data on cost burden by owners}
\usage{
process_burden_owners(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on cost burden by owners}
}
\value{
Processed data on cost burden for owners
}
\description{
Process cost-burden for homeowners
Processes ACS5 tidycensus downloaded data on cost burden by owners
}
