% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_burden_owners_mortgage}
\alias{process_burden_owners_mortgage}
\title{Process cost-burden for owners by mortgage status}
\usage{
process_burden_owners_mortgage(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on cost-burden for owners by mortgage status}
}
\value{
Processed data on cost-burden for owners by mortgage status
}
\description{
Processes ACS5 tidycensus downloaded data on cost-burden for owners by mortgage status
}
