% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_burden_renters}
\alias{process_burden_renters}
\title{Process cost-burden for renters
Processes ACS5 tidycensus downloaded data on cost-burden for renters}
\usage{
process_burden_renters(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on cost-burden for renters}
}
\value{
Processed data on cost-burden for renters
}
\description{
Process cost-burden for renters
Processes ACS5 tidycensus downloaded data on cost-burden for renters
}
