% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_computer_age}
\alias{process_computer_age}
\title{Process Computer/Internet by Age
Processes ACS5 tidycensus downloaded data on computer/internet usage by age}
\usage{
process_computer_age(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on computer/internet usage by age}
}
\value{
Processed data on computer/internet usage by age
}
\description{
Process Computer/Internet by Age
Processes ACS5 tidycensus downloaded data on computer/internet usage by age
}
