% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_computer_overall}
\alias{process_computer_overall}
\title{Process Computer/Internet Data}
\usage{
process_computer_overall(df)
}
\arguments{
\item{df}{tidycensus downloaded data on computer/internet usage}
}
\value{
processed data on computer/internet usage
}
\description{
Processes tidycensus downloaded data on computer/internet usage
}
