% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_test.R
\name{process_df}
\alias{process_df}
\title{Process ACS data
Groups lightly-processed tidycensus/ACS dataframe, and identifies statistical-significance of values comparing Takoma Park to Montgomery County and Maryland.}
\usage{
process_df(
  df,
  group_cols,
  overall_cols,
  name_col,
  bind_overall = NULL,
  root_df = NULL
)
}
\arguments{
\item{df}{Lightly-processed ACS dataframe with estimate and MOE column.}

\item{group_cols}{Columns to group ACS dataframe by.}

\item{overall_cols}{The set of columns to calculate overall results for. E.g., if you want to see how age groups differ from overall rates of access, you would enter the computer access column as the group col.}

\item{name_col}{Base-name of column to store values in for grouped dataframe.}

\item{bind_overall}{Whether to add overall results as a row to the processed dataframe. Should enter name of column with groups comparing against (e.g., if comparing ages against overall, should enter age).}

\item{root_df}{If the df has already been processed more, the base-dataframe from which the df was processed.}
}
\value{

}
\description{
Process ACS data
Groups lightly-processed tidycensus/ACS dataframe, and identifies statistical-significance of values comparing Takoma Park to Montgomery County and Maryland.
}
