% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_disab_age}
\alias{process_disab_age}
\title{Process disability by age}
\usage{
process_disab_age(df, disab_age_df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on disability by age}

\item{disab_age_df}{Disability by age data}
}
\value{
Processed disability by age data
}
\description{
Processes ACS5 tidycensus downloaded data on disability by age
}
