% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_disab_foodstamp}
\alias{process_disab_foodstamp}
\title{Process foodstamp and disability status data}
\usage{
process_disab_foodstamp(df, overall = F)
}
\arguments{
\item{df}{tidycensus-downloaded data on household food-stamp receipt by disability status, census table B22010}

\item{overall}{Whether to aggregate dataframe to households by presence of person with a disability; default F}
}
\value{
Processed dataframe on household food-stamp receipt by disability status, or households by presence of person with a disability
}
\description{
Processes tidycensus-downloaded data from table B22010 on household food-stamp receipt by disability status. Option to calculate overall-results for household disability status, since other Census tables on disability are not available at the block-group level
}
