% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_disab_sex}
\alias{process_disab_sex}
\title{Process disability by sex
Processes ACS5 tidycensus downloaded data on disability by sex}
\usage{
process_disab_sex(df, df_disab_sex)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on disability}

\item{df_disab_sex}{ACS5 tidycensus downloaded data on disability by sex}
}
\value{
Processed disability by sex data
}
\description{
Process disability by sex
Processes ACS5 tidycensus downloaded data on disability by sex
}
