% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_disab_type}
\alias{process_disab_type}
\title{Process disability by type
Processes ACS5 tidycensus downloaded data on disability by type}
\usage{
process_disab_type(df, df_disab_pop)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on disability by type}

\item{df_disab_pop}{Data on disability as percent of population}
}
\value{

}
\description{
Process disability by type
Processes ACS5 tidycensus downloaded data on disability by type
}
