% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_disability_overall}
\alias{process_disability_overall}
\title{Process disability overall}
\usage{
process_disability_overall(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on disability overall with disability column}
}
\value{
Processed data on disability overall with disability column
}
\description{
Processes ACS5 tidycensus downloaded data on disability overall
}
