% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_disability_pop}
\alias{process_disability_pop}
\title{Process disability as percent of population
Processes ACS5 tidycensus downloaded data on disability, calculating percentages of place's population}
\usage{
process_disability_pop(df_disability)
}
\arguments{
\item{df_disability}{ACS5 tidycensus downloaded data on percentages of people with a disability}
}
\value{

}
\description{
Process disability as percent of population
Processes ACS5 tidycensus downloaded data on disability, calculating percentages of place's population
}
