% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_educ_overall}
\alias{process_educ_overall}
\title{Process education data
Processes ACS5 tidycensus downloaded data on educational attainment}
\usage{
process_educ_overall(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on educational attainment}
}
\value{
Processed data on educational attainment
}
\description{
Process education data
Processes ACS5 tidycensus downloaded data on educational attainment
}
