% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_educ_race}
\alias{process_educ_race}
\title{Process educational attainment by race}
\usage{
process_educ_race(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on educational attainment by race}
}
\value{
Processed data on educational attainment by race
}
\description{
ACS5 tidycensus downloaded data on educational attainment by race
}
