% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_educ_race_gender}
\alias{process_educ_race_gender}
\title{Process education/race/sex data}
\usage{
process_educ_race_gender(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on educational attainment by race and sex}
}
\value{
Processed data on educational attainment by race and sex
}
\description{
Process education/race/sex data
}
