% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_employment_age}
\alias{process_employment_age}
\title{Process employment/age data}
\usage{
process_employment_age(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on employment by age}
}
\value{
Processed data on employment by age
}
\description{
Processes ACS5 tidycensus downloaded data on employment by age
}
