% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_employment_age_gender}
\alias{process_employment_age_gender}
\title{Process Employment/Age/Gender Data}
\usage{
process_employment_age_gender(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on employment by age and gender}
}
\value{
Processed data on employment by age and gender
}
\description{
Processes ACS5 tidycensus downloaded data on employment by age and gender
}
