% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_employment_age_gender_base}
\alias{process_employment_age_gender_base}
\title{Base-process employment/age/gender
Base-processing of ACS5 tidycensus downloaded data on employment/age/gender}
\usage{
process_employment_age_gender_base(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on employment/age/gender}
}
\value{
Base-processed ACS5 tidycensus downloaded data on employment/age/gender
}
\description{
Base-process employment/age/gender
Base-processing of ACS5 tidycensus downloaded data on employment/age/gender
}
