% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_family}
\alias{process_family}
\title{Process family-status data
Processes tidycensus downloaded ACS5 data on family status}
\usage{
process_family(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on family}
}
\value{
Processed data on family status
}
\description{
Process family-status data
Processes tidycensus downloaded ACS5 data on family status
}
