% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_family_poverty}
\alias{process_family_poverty}
\title{Process poverty by family status
Processes tidycensus downloaded ACS5 data on poverty by family status}
\usage{
process_family_poverty(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on poverty by family status}
}
\value{
Processed data on poverty by family status
}
\description{
Process poverty by family status
Processes tidycensus downloaded ACS5 data on poverty by family status
}
