% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_family_poverty_race}
\alias{process_family_poverty_race}
\title{Process family by poverty and race
Processes tidycensus downloaded ACS5 data on family-status by poverty and race}
\usage{
process_family_poverty_race(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on family-status by poverty and race}
}
\value{

}
\description{
Process family by poverty and race
Processes tidycensus downloaded ACS5 data on family-status by poverty and race
}
