% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_family_pubassist}
\alias{process_family_pubassist}
\title{Process family-type by public assistance}
\usage{
process_family_pubassist(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on family-type by receipt of public assistance}
}
\value{
tidycensus downloaded ACS5 data on family-type by public assistance
}
\description{
Process family-type by public assistance
}
