% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_family_type_income}
\alias{process_family_type_income}
\title{Process family type by income
Processes tidycensus downloaded ACS5 data on family type by income}
\usage{
process_family_type_income(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on family type by income}
}
\value{
Processed data on family type by income
}
\description{
Process family type by income
Processes tidycensus downloaded ACS5 data on family type by income
}
