% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_famtype_incomelev}
\alias{process_famtype_incomelev}
\title{Process tidycensus ACS5 data on Family type by Income-level}
\usage{
process_famtype_incomelev(df)
}
\arguments{
\item{df}{Tidycensus downloaded dataframe on family type by income level (table B19131), with tidycensus downloaded variables dataframe joined to it and renamed to lower}
}
\value{
Processed dataframe on family type by income level
}
\description{
Processes tidycensus dowloaded data on family type by income level; table B19131
}
