% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_foreign_born}
\alias{process_foreign_born}
\title{Process foreign born}
\usage{
process_foreign_born(df)
}
\arguments{
\item{df}{tidycensus downloaded data on birthplace}
}
\value{
processed dataframe on birth by location
}
\description{
Processes tidycensus downloaded data on birthplace
}
