% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_foreign_born_birth}
\alias{process_foreign_born_birth}
\title{Process foreign born birth}
\usage{
process_foreign_born_birth(df)
}
\arguments{
\item{df}{tidycensus downloaded data on country of origin}
}
\value{
processed tidycensus dataset with country of origin results by continent
}
\description{
Processes tidycensus downloaded data on birth by country of origin
}
