% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_health_age}
\alias{process_health_age}
\title{Process health-insurance by age
Processes ACS5 tidycensus downloaded data on health insurance by age}
\usage{
process_health_age(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on health insurance by age}
}
\value{
Processed health-insurance by age data
}
\description{
Process health-insurance by age
Processes ACS5 tidycensus downloaded data on health insurance by age
}
