% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_health_race}
\alias{process_health_race}
\title{Process health insurance by race
Processes ACS5 tidycensus downloaded data on health insurance by race}
\usage{
process_health_race(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on health insurance by race}
}
\value{
Processed data on health insurance by race
}
\description{
Process health insurance by race
Processes ACS5 tidycensus downloaded data on health insurance by race
}
