% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_home_lang}
\alias{process_home_lang}
\title{Process language spoken at home}
\usage{
process_home_lang(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on language spoken at home}
}
\value{
Processed data on language spoken at home
}
\description{
Processes ACS5 tidycensus downloaded data on language spoken at home
}
