% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_hous_famtyp}
\alias{process_hous_famtyp}
\title{Process ACS-5 year tidycensus data on household by family type}
\usage{
process_hous_famtyp(df)
}
\arguments{
\item{df}{Base-loaded dataframe on housholds by family type at a given geography, renamed to lower with tidycensus variables joined to it}
}
\value{
Processed dataframe on households by family type
}
\description{
Processes ACS-5 year tidycensus data on household by family type; table B11001.
}
