% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_hous_povage}
\alias{process_hous_povage}
\title{Process households by age and poverty-status}
\usage{
process_hous_povage(df)
}
\arguments{
\item{df}{Tidycensus-downloaded data on age by poverty-status, table B17017}
}
\value{
Processed data on age by poverty-status
}
\description{
Processes tidycensus downloaded data on households by age and poverty-status, table B17017
}
