% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_household_income_ownercosts}
\alias{process_household_income_ownercosts}
\title{Process cost burden by household income for owners
Processes ACS5 tidycensus downloaded data on cost-burden for owners by household income}
\usage{
process_household_income_ownercosts(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on cost-burden for owners by household income}
}
\value{
Processed cost-burden by household income data for owners
}
\description{
Process cost burden by household income for owners
Processes ACS5 tidycensus downloaded data on cost-burden for owners by household income
}
