% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_household_income_rentercosts}
\alias{process_household_income_rentercosts}
\title{Process cost-burden by household income for renters}
\usage{
process_household_income_rentercosts(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on cost-burden by household income for renters}
}
\value{
Processed data on cost burden by household income for renters
}
\description{
Processes ACS5 tidycensus downloaded data on cost-burden by household income for renters
}
