% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_income_last12}
\alias{process_income_last12}
\title{Process income data
Processes ACS5 tidycensus downloaded data on income in last 12 months}
\usage{
process_income_last12(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on income in last 12 months}
}
\value{
Processed data on income in last 12 months
}
\description{
Process income data
Processes ACS5 tidycensus downloaded data on income in last 12 months
}
