% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_income_race}
\alias{process_income_race}
\title{Process race/household income}
\usage{
process_income_race(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on race and household income}
}
\value{
Processed data on race/household income
}
\description{
Processes ACS5 tidycensus downloaded data on race and household income
}
