% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_income_race_median}
\alias{process_income_race_median}
\title{Process median income by race
Processes ACS5 tidycensus downloaded data on household median income by race}
\usage{
process_income_race_median(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on household median income by race}
}
\value{
Processed data on median income by race
}
\description{
Process median income by race
Processes ACS5 tidycensus downloaded data on household median income by race
}
