% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_incpov_un1un2}
\alias{process_incpov_un1un2}
\title{Process income-poverty df to under 1 and under 2x ratio}
\usage{
process_incpov_un1un2(df_povratio)
}
\arguments{
\item{df_povratio}{tidycensus downloaded income-poverty ratio dataset}
}
\value{
Income-poverty ratio dataset with totals of under the poverty level and 2 times the poverty level
}
\description{
Processes tidycensus downloaded income-poverty ratio dataset to show under 1x the poverty level and under 2x the poverty level
}
