% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_own_costburden_age}
\alias{process_own_costburden_age}
\title{Process owner cost-burden by age}
\usage{
process_own_costburden_age(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on owner cost-burden by age}
}
\value{
Processed data on owner cost-burden by age
}
\description{
Processes ACS5 tidycensus downloaded data on owner cost-burden by age
}
