% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_owner_income_num}
\alias{process_owner_income_num}
\title{Process income for owners
Processes ACS5 tidycensus downloaded data on income for owners}
\usage{
process_owner_income_num(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on income for owners}
}
\value{
Processed data on income for owners
}
\description{
Process income for owners
Processes ACS5 tidycensus downloaded data on income for owners
}
