% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_pop_tot}
\alias{process_pop_tot}
\title{Population totals from race/ethnicity}
\usage{
process_pop_tot(race_ethn_processed_df)
}
\arguments{
\item{race_ethn_processed_df}{Processed race/ethnicity data}
}
\value{
Dataframe with population totals for given place
}
\description{
Grab population totals from processed race/ethnciity data
}
