% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_poverty_race_age}
\alias{process_poverty_race_age}
\title{Process poverty by race and age data
Processes tidycensus downloaded ACS5 data on poverty by race by age}
\usage{
process_poverty_race_age(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on poverty by race by age}
}
\value{
Processes data on poverty by race by age
}
\description{
Process poverty by race and age data
Processes tidycensus downloaded ACS5 data on poverty by race by age
}
