% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_poverty_sex}
\alias{process_poverty_sex}
\title{Process poverty by sex data}
\usage{
process_poverty_sex(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on poverty by sex}
}
\value{
Processed data on poverty by sex
}
\description{
Processes tidycensus downloaded ACS5 data on poverty by sex
}
