% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_poverty_sex_age}
\alias{process_poverty_sex_age}
\title{Process poverty by sex and age}
\usage{
process_poverty_sex_age(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on poverty by sex and age}
}
\value{
Processed data on poverty by sex and age
}
\description{
Processes tidycensus downloaded ACS5 data on poverty by sex and age
}
