% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_age_gender}
\alias{process_race_age_gender}
\title{Process race, age, gender
Processes tidycensus downloaded race, age, and gender data}
\usage{
process_race_age_gender(df)
}
\arguments{
\item{df}{tidycensus downloaded race, age, and gender data}
}
\value{
Processed race, age, and gender data
}
\description{
Process race, age, gender
Processes tidycensus downloaded race, age, and gender data
}
