% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_assist}
\alias{process_race_assist}
\title{Process race by public assistance data
Processes tidycensus downloaded ACS5 data on receipt of public assistance by race}
\usage{
process_race_assist(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on receipt of public assistance by race}
}
\value{
Processed data on receipt of public assistance by race
}
\description{
Process race by public assistance data
Processes tidycensus downloaded ACS5 data on receipt of public assistance by race
}
