% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_computer}
\alias{process_race_computer}
\title{Process Computer Usage by Race}
\usage{
process_race_computer(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on computer/internet usage by race}
}
\value{
processed data on computer/internet usage by race
}
\description{
Processes ACS5 tidycensus downloaded data on computer/internet usage by race
}
