% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_disability}
\alias{process_race_disability}
\title{Process race/disability data}
\usage{
process_race_disability(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on disability and race and age}
}
\value{
tidycensus downloaded ACS5 data on race/disability
}
\description{
Processes tidycensus downloaded ACS5 data on disability by race
}
