% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_disability_age}
\alias{process_race_disability_age}
\title{Pull race/disability/age}
\usage{
process_race_disability_age(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on race/disability/age}
}
\value{
Dataframe with variables for age, disability, race
}
\description{
Pulls race/disability/age from tidycensus downloaded ACS5 data
}
