% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_employ_gender_u65}
\alias{process_race_employ_gender_u65}
\title{Process Race/Employment/Gender under 65}
\usage{
process_race_employ_gender_u65(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on employment by race and gender for people under 65}
}
\value{
Processed data on employment by race and gender for people under 65
}
\description{
Processes ACS5 tidycensus downloaded data on employment by race and gender for people under 65
}
