% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_employment_age_gender_base}
\alias{process_race_employment_age_gender_base}
\title{Base-processing employment by race, age, and gender}
\usage{
process_race_employment_age_gender_base(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on employment by race, age, and gender}
}
\value{
Processed data on employment by race, age, and gender
}
\description{
Base processing of ACS5 tidycensus downloaded data on employment by race, age, and gender
}
