% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_employment_u65_base}
\alias{process_race_employment_u65_base}
\title{Base-processing of race/employment under 65 data}
\usage{
process_race_employment_u65_base(df, tot_df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on employment by race and gender for people under 65}

\item{tot_df}{Totals dataset}
}
\value{
Base-processed data on employment by race and gender under 65
}
\description{
Performs base-processing of race/employment under 65 data
}
