% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_ethn}
\alias{process_race_ethn}
\title{Process race and ethnicity data
Processes tidycensus downloaded race/ethnicity data}
\usage{
process_race_ethn(df)
}
\arguments{
\item{df}{tidycensus downloaded race/ethnicity data file}
}
\value{
processed race/ethnicity data
}
\description{
Process race and ethnicity data
Processes tidycensus downloaded race/ethnicity data
}
