% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_foreign_born}
\alias{process_race_foreign_born}
\title{Process foreign born by race}
\usage{
process_race_foreign_born(df)
}
\arguments{
\item{df}{tidycensus downloaded dataset of birthplace by race}
}
\value{
Processed tidycensus dataframe of birthplace by race
}
\description{
Process foreign born by race
}
