% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_gender_employment_general}
\alias{process_race_gender_employment_general}
\title{Process employment by race and gender}
\usage{
process_race_gender_employment_general(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on employment by race and gender}
}
\value{
Processed data on employment by race and gender
}
\description{
Process ACS5 tidycensus downloaded data on employment by race and gender
}
