% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_home_own}
\alias{process_race_home_own}
\title{Base-process home-ownership by race}
\usage{
process_race_home_own(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on home-ownership by race}
}
\value{
Processed data on home-ownership by race
}
\description{
ACS5 tidycensus downloaded data on home-ownership by race
}
