% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_home_own_race_tenure}
\alias{process_race_home_own_race_tenure}
\title{Process home-ownership by race and tenure}
\usage{
process_race_home_own_race_tenure(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on tenure by race}
}
\value{
Processed data on tenure by race
}
\description{
Processes ACS5 tidycensus downloaded data on tenure and race
}
