% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_internet_base}
\alias{process_race_internet_base}
\title{Process internet/race data
Processes ACS5 tidycensus downloaded data on internet/race, in preparation for other processing}
\usage{
process_race_internet_base(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on internet usage by race}
}
\value{
Base-level processed data on internet usage by race
}
\description{
Process internet/race data
Processes ACS5 tidycensus downloaded data on internet/race, in preparation for other processing
}
