% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_internet_complete}
\alias{process_race_internet_complete}
\title{Process race/internet data}
\usage{
process_race_internet_complete(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on race/internet usage}
}
\value{
Processed data on race/internet usage
}
\description{
Processes ACS5 tidycensus downloaded data on race/internet usage
}
