% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_laborforce_gender_u65}
\alias{process_race_laborforce_gender_u65}
\title{Process labor force by race and gender for people under 65}
\usage{
process_race_laborforce_gender_u65(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on labor force participation by race and gender for people under 65}
}
\value{
Processed data on employment by race and gender for people under 65
}
\description{
Process labor force by race and gender for people under 65
}
