% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_race_transport_complete}
\alias{process_race_transport_complete}
\title{Process transport to work by race}
\usage{
process_race_transport_complete(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on transport to work by race}
}
\value{
Processed data on transport to work by race
}
\description{
Processes tidycensus downloaded ACS5 data on transport to work by race
}
