% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_read_disab_type}
\alias{process_read_disab_type}
\title{Processes disability by type
Processes ACS5 tidycensus downloaded data on disability by type}
\usage{
process_read_disab_type(df, pop_tot_df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on disability by type}

\item{pop_tot_df}{Dataframe with population totals}
}
\value{
Processed data on disability by type
}
\description{
Processes disability by type
Processes ACS5 tidycensus downloaded data on disability by type
}
