% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_rent_own_costburden_age}
\alias{process_rent_own_costburden_age}
\title{Process owner and renter cost-burden by age}
\usage{
process_rent_own_costburden_age(rent_df, own_df)
}
\arguments{
\item{rent_df}{ACS5 tidycensus downloaded data on renter cost-burden by age}

\item{own_df}{ACS5 tidycensus downloaded data on owner cost-burden by age}
}
\value{
Lightly processed data on renter/owner cost burden by age
}
\description{
Processes ACS5 tidycensus downloaded datasets on renter cost-burden by age, and owner cost-burden by age, and binds them together
}
