% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_renter_income_num}
\alias{process_renter_income_num}
\title{Process income for renters
Processes ACS5 tidycensus downloaded data on income for renters}
\usage{
process_renter_income_num(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on income for renters}
}
\value{
Processed data on income for renters
}
\description{
Process income for renters
Processes ACS5 tidycensus downloaded data on income for renters
}
