% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_tenure_age}
\alias{process_tenure_age}
\title{Process tenure by age
Processes ACS5 tidycensus downloaded data on tenure by age}
\usage{
process_tenure_age(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on tenure by age}
}
\value{
Processed data on tenure by age
}
\description{
Process tenure by age
Processes ACS5 tidycensus downloaded data on tenure by age
}
