% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_tenure_df}
\alias{process_tenure_df}
\title{Process tenure}
\usage{
process_tenure_df(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on tenure}
}
\value{
Processed data on tenure
}
\description{
Processes ACS5 tidycensus downloaded data on tenure
}
