% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_tenure_income_median}
\alias{process_tenure_income_median}
\title{Process median income by tenure}
\usage{
process_tenure_income_median(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on median income by tenure}
}
\value{
Processed data on median income by tenure
}
\description{
Processes ACS5 tidycensus downloaded data on median income by tenure
}
