% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_tenure_race}
\alias{process_tenure_race}
\title{Process tenure by race
Processes ACS5 tidycensus downloaded data on tenure by race}
\usage{
process_tenure_race(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on tenure by race}
}
\value{
Processed data on tenure by race
}
\description{
Process tenure by race
Processes ACS5 tidycensus downloaded data on tenure by race
}
