% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_tenure_vehicleown}
\alias{process_tenure_vehicleown}
\title{Process vehicle-ownership data by tenure}
\usage{
process_tenure_vehicleown(df, overall = F)
}
\arguments{
\item{df}{Dataframe on number of vehicles owned by tenure, reflecting table B25044 in 2020 5-year ACS}

\item{overall}{Whether to aggregate data up to level of owning a vehicle (without tenure); default FALSE}
}
\value{
Processed-dataframe on vehicle ownership by tenure, or households by vehicle ownership
}
\description{
Processes tidycensus-downloaded tenure by number of vehicles owned data, table B25044 in 2020 5-year ACS
}
