% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_vetage}
\alias{process_vetage}
\title{Process veteran-status by sex and age dataset into veteran-status by age}
\usage{
process_vetage(df, bindoverall = "age")
}
\arguments{
\item{df}{Unprocessed ACS5 veteran-status by sex and age dataset}

\item{bindoverall}{Default "age"; set to NULL to not add an extra row to the dataset containing results for veteran-status for the overall population}
}
\value{
Processed datset of veteran-status by sex and age
}
\description{
Processes unprocessed veteran-status by sex and age dataset, turning it into a dataset of veteran-status by age
}
