% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{puller_funct}
\alias{puller_funct}
\title{Prepare ACS data
Function to standardize tidycensus downloaded ACS data, with option to filter to a specific place}
\usage{
puller_funct(
  df,
  place_string = "",
  no_pull = FALSE,
  variables = acsprocess::variables_2019
)
}
\arguments{
\item{df}{Downloaded tidycensus ACS dataframe}

\item{place_string}{String name of tidycensus place to filter data to}

\item{no_pull}{True/false whether to filter dataframe to tidycensus place}

\item{variables}{tidycensus dataframe of variables; default 2019}
}
\value{
tidycensus dataframe with column names in lower case and variable labels associated with tidycensus data
}
\description{
Prepare ACS data
Function to standardize tidycensus downloaded ACS data, with option to filter to a specific place
}
