% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_load.R
\name{quickdircreate}
\alias{quickdircreate}
\title{Check if directory exists and create it if not}
\usage{
quickdircreate(dirpath)
}
\arguments{
\item{dirpath}{Path to directory checking exists and creating if not}
}
\value{
Created directory
}
\description{
Checks if directory exists and create it if not
}
