% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_functs.R
\name{race_factor}
\alias{race_factor}
\title{Race factor
Orders race column in given data-frame for consistency in plot_ly coloring, based on what races present in dataframe}
\usage{
race_factor(df, race_col = race)
}
\arguments{
\item{df}{Dataframe with race column}

\item{race_col}{Name of race column}
}
\value{
Dataframe with race column as factor
}
\description{
Race factor
Orders race column in given data-frame for consistency in plot_ly coloring, based on what races present in dataframe
}
