% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_pull.R
\name{race_pull}
\alias{race_pull}
\title{Pull race values as column from tidycensus DF}
\usage{
race_pull(df, race_col = "race", col = "concept")
}
\arguments{
\item{df}{A tidycensus dataframe with race values stored in a string in a column}

\item{race_col}{The name of the column to store race values in}

\item{col}{A column from which to extract race values, usually concept}
}
\value{
A tidycensus dataframe with race values stored in a column
}
\description{
Pull race values as column from tidycensus DF
}
