% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_functs.R
\name{race_recode}
\alias{race_recode}
\title{Recode ACS race  values}
\usage{
race_recode(df, race_col, new_race_col = "race")
}
\arguments{
\item{df}{A dataframe with tidycensus ACS race values}

\item{race_col}{The column in the dataframe containing tidycensus ACS race values}

\item{new_race_col}{The name of the new column with renamed race values as an ordered factor based on what race values are present in the dataframe}
}
\value{
A dataframe with a renamed ACS race column
}
\description{
Function to recode a tidycensus ACS dataframe with a race column to shorter labels. This is best used after the race_pull function, which separates the tidycensus dataframe "concept" column into a meaningful race value.
}
