% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_recode.R
\name{recode_tenure}
\alias{recode_tenure}
\title{Recode tenure
Extracts tenure values from ACS dataframe}
\usage{
recode_tenure(df, tenure_col)
}
\arguments{
\item{df}{Dataframe with tenure column}

\item{tenure_col}{Name of tenure column}
}
\value{
Dataframe with tenure column
}
\description{
Recode tenure
Extracts tenure values from ACS dataframe
}
