% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_label.R
\name{separate_label}
\alias{separate_label}
\title{Separate tidycensus label column}
\usage{
separate_label(df, names_vector, label_col = "label")
}
\arguments{
\item{df}{a tidycensus loaded dataset}

\item{names_vector}{a vector of new column names corresponding to the number of variables created by the separate function. Can enter NA to omit}

\item{label_col}{the name of the label column; usually label}
}
\value{
a tidycensus dataset with new variables corresponding to the split string
}
\description{
tidycensus datasets come with a label function containing information on the row of interest in one string. This function separates the string into multiple columns defined by the names_vector
}
