% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_test.R
\name{signif_checker}
\alias{signif_checker}
\title{Significance check
Loops through named list of columns containing statistical-significance results, and creates new column describing any statistically significant results.}
\usage{
signif_checker(
  data,
  signif_cols = list(Overall = "signif_overall", MC = "signif_mont", MD =
    "signif_maryland")
)
}
\arguments{
\item{data}{Dataframe with columns containing descriptions of significantly significant results against other comparison groups (e.g., another location, overall vs. a group).}

\item{signif_cols}{Named list with value representing name of the column with significance values, and names representing how significance should be described in the new column (e.g., "Montgomery County" = "signif_mont" would be for a signif_mont column describing the significance of differences with Montgomery County, and Montgomery County is how it would be represented in the new column),}
}
\value{
A dataframe with a column describing all statistically significant differences with comparison groups (e.g., Montgomery County, Maryland, and Overall).
}
\description{
Significance check
Loops through named list of columns containing statistical-significance results, and creates new column describing any statistically significant results.
}
