% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_test.R
\name{signif_test}
\alias{signif_test}
\title{Significance test for ACS estimates}
\usage{
signif_test(est1, est2, moe1, moe2, just_val = F)
}
\arguments{
\item{est1}{First estimate to compare}

\item{est2}{Second estimate to compare}

\item{moe1}{MOE of first estimate}

\item{moe2}{MOE of second estimate}

\item{just_val}{Whether to return just the level of confidence or a list with the critical value and level of confidence}
}
\value{
A single value of whether/at what level of confidence the estimates differ statistically, or a list with the critical value and first value
}
\description{
Significance test for ACS estimates
}
