% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_col_add.R
\name{total_col_add}
\alias{total_col_add}
\title{Create tidycensus total column}
\usage{
total_col_add(df, total_cols, join_col = "name", est_col = "estimate")
}
\arguments{
\item{df}{a tidycensus dataframe with a total represented as one row}

\item{total_cols}{a list, formatted as list(name of new column = name column to pull total for). If you've separated the label column, the name of column to pull for will be NA.}

\item{join_col}{a column or multiple columns to join the totals to. Often name, unless a race-disaggregated dataset without an aggregate total across races (e.g., race by age). If multiple totals are being calculated, then the columns are joined in sequence (e.g., pulling totals for race, then race by age, then race by age by gender). If the there is one total column and multiple join columns, the columns are joined all at once (e.g., to generate unique totals for a place and race). In situations involving a sequence where a unique observation represents a combination of two variables, it's best to merge the two variables in advance to one column.}

\item{est_col}{the name of the column containing values for each group; unless renamed or transformed, estimate}
}
\value{
a tidycensus dataset with totals for different groups as columns
}
\description{
A function to pull one or multiple total values in a tidycensus dataset for a group or groups from a row and make it into a column. This is best used after separate_label; the group total will be NA after the label is separated, so you enter the name of the group column as the total_col value e.g., ("pop_total" = race_ethnicity) to create a column for the population total in a dataset where race_ethnicity has the racial/ethnic group for each row.
}
