% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disab_pull.R, R/process_functions.R
\name{disab_pull}
\alias{disab_pull}
\title{Pull disability values from concept column}
\usage{
disab_pull(df, new_col = disability, concept_col = concept)

disab_pull(df, new_col = disability, concept_col = concept)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on disability by type}

\item{new_col}{Name of new column that will store disability-type, data-masked; default disability}

\item{concept_col}{Data-masked name of concept column; default concept}
}
\value{


ACS5 tidycensus downloaded data on disability by type with disability column
}
\description{
Creates column on disability by type from ACS5 tidycensus downloaded data on disability by type
}
