% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_test.R
\name{process_df}
\alias{process_df}
\title{Process ACS data}
\usage{
process_df(
  df,
  group_cols,
  overall_cols,
  name_col,
  bind_overall = NULL,
  root_df = NULL
)
}
\arguments{
\item{df}{Lightly-processed ACS dataframe with estimate and MOE column.}

\item{group_cols}{String vector of columns to group ACS dataframe by. E.g., if the dataset contains information on veteran status by age and sex, and you want a dataframe with information on veteran status by age, you'd enter c("age", "veteran").}

\item{overall_cols}{String vector of columns to calculate overall results for. E.g., if you want to see how computer access by age groups differ from overall rates of computer access, you would enter the computer access column as the group col.}

\item{name_col}{String base-name of column to store values in for grouped dataframe.}

\item{bind_overall}{Whether to add overall results as a row to the processed dataframe. Default NULL. To add an overall row, should enter the string name of the column containing groups comparing against (e.g., if comparing ages against overall, should enter age; in the new row, the age column will have a value of "Overall").}

\item{root_df}{If the df has already been processed more, the base-dataframe from which the df was processed.}
}
\value{
Dataframe grouped by group columns, with statistical significance of differences from overall_cols.
}
\description{
Groups lightly-processed tidycensus/ACS dataframe, and identifies statistical-significance of values comparing subgroups to an overall group.
}
