% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_disability_age}
\alias{process_disability_age}
\title{Process disability by age}
\usage{
process_disability_age(df_disability)
}
\arguments{
\item{df_disability}{ACS5 tidycensus downloaded data on disability by age}
}
\value{
Processed data on disability by age
}
\description{
Processes ACS5 tidycensus downloaded data on rates of disability by age
}
