% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_family_tenure}
\alias{process_family_tenure}
\title{Process family-type by tenure}
\usage{
process_family_tenure(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on family-type by tenure}
}
\value{
tidycensus downloaded ACS5 data on family-type by tenure
}
\description{
Processes tidycensus downloaded ACS5 data on family-type by tenure data
}
