% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_gender}
\alias{process_gender}
\title{Process sex data
Processes tidycensus downloaded ACS5 data on sex}
\usage{
process_gender(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on sex}
}
\value{
Processed data on sex
}
\description{
Process sex data
Processes tidycensus downloaded ACS5 data on sex
}
