% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_poverty}
\alias{process_poverty}
\title{Process poverty data
Processes tidycensus downloaded ACS5 data on poverty}
\usage{
process_poverty(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on poverty}
}
\value{
Processed data on poverty
}
\description{
Process poverty data
Processes tidycensus downloaded ACS5 data on poverty
}
