% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_poverty_detail}
\alias{process_poverty_detail}
\title{Process detailed poverty data}
\usage{
process_poverty_detail(df)
}
\arguments{
\item{df}{tidycensus downloaded ACS5 data on poverty}
}
\value{
ACS5 data on poverty
}
\description{
Processes detailed tidycensus downloaded ACS5 data on poverty
}
