% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_sex_disab_overall}
\alias{process_sex_disab_overall}
\title{Process disability by sex}
\usage{
process_sex_disab_overall(df)
}
\arguments{
\item{df}{ACS5 tidycensus downloaded data on disability by sex}
}
\value{
Processed disability by sex data
}
\description{
Processes ACS5 tidycensus downloaded data on rates of disability by sex
}
